Add files via upload
The image presents a simulated satellite view of the Iberian Peninsula, as it might appear in the year 2125. Based on one possible climate-change modelling scenario, much of the southern portion of the Peninsula has become desert, reflecting a northward expansion of the Sahara. Smaller desertified areas also appear along the Mediterranean coast of Spain and on the island of Ibiza. At the bottom of the image, parts of northern Africa are visible, and they too appear as desert terrain. The simulation is adapted from data presented by Natalie Thomas and Sumant Nigam in the article “Twentieth-century climate change over Africa: Seasonal hydroclimate trends and Sahara desert expansion” (Journal of Climate 31.9, 2018, pp. 3349–3370). A text column with the title, description, and technical details is displayed on the left side of the image.